+++ src/pages/register/confirm.jsx
@@ -1,7 +1,7 @@
 import { useState, createRef, useEffect } from 'react'
-import Progress from '../components/Progress'
-import Watch from './Timer'
-import { useLocalStorage } from '../utils/useLocalStorage'
+import Progress from '../../components/Progress'
+import Watch from '../Timer'
+import { useLocalStorage } from '../../utils/useLocalStorage'
 import axios from 'axios'
 import toast from 'react-hot-toast'
 import { useRouter } from 'next/router'
@@ -83,7 +83,7 @@ export default function confirm(props) {
           ))
           setTimeout(() => {
             localStorage.clear()
-            router.push('/terms')
+            router.push('/register/terms')
           }, 2000)
         } else {
           toast.custom((t) => (
