+++ src/pages/register/phone.jsx
@@ -1,11 +1,11 @@
 import * as yup from 'yup'
 import { useFormik } from 'formik'
-import Button from '../components/Button'
+import Button from '../../components/Button'
 import TextField from '@material-ui/core/TextField'
 import InputMask from 'react-input-mask'
 import { useRouter } from 'next/router'
-import { useLocalStorage } from '../utils/useLocalStorage'
-import Progress from '../components/Progress'
+import { useLocalStorage } from '../../utils/useLocalStorage'
+import Progress from '../../components/Progress'
 
 export default function phone() {
   const router = useRouter()
@@ -20,7 +20,7 @@ export default function phone() {
     onSubmit: (values) => {
       localStorage.setItem('type', 1)
       localStorage.setItem('mobile_phone', JSON.stringify(values.celular))
-      router.push('/password')
+      router.push('/register/password')
     },
   })
