+++ src/pages/register/terms.tsx
@@ -1,12 +1,12 @@
 import Image from 'next/image'
 import { useRouter } from 'next/router'
-import ConcluirImg from '../../public/undrawn.png'
-import Progress from '../components/Progress'
+import ConcluirImg from '../../../public/undrawn.png'
+import Progress from '../../components/Progress'
 
 export default function terms() {
   const router = useRouter()
   function handleClick() {
-    router.push('/conclusion')
+    router.push('/register/conclusion')
   }
   return (
     <>
